hotpotqa/requirements.txt
openai
google-genai
gymnasium
requests
beautifulsoup4
numpy
pandas
matplotlib
pytest
